+++ Rebus.Tests/Serialization/TestJsonSerializer.cs
@@ -6,7 +6,10 @@ using System.Threading;
 using System.Threading.Tasks;
 using Newtonsoft.Json;
 using NUnit.Framework;
+using Rebus.Extensions;
 using Rebus.Messages;
+using Rebus.Serialization;
+using Rebus.Serialization.Custom;
 using Rebus.Tests.Contracts;
 using JsonSerializer = Rebus.Serialization.Json.JsonSerializer;
 #pragma warning disable 4014
@@ -20,23 +23,31 @@ namespace Rebus.Tests.Serialization
 
         protected override void SetUp()
         {
-            _serializer = new JsonSerializer();
+            _serializer = new JsonSerializer(new SimpleAssemblyQualifiedMessageTypeNameConvention());
         }
 
         /*
          Initial:
-Made 23133 iterations in 5s
-Made 81106 iterations in 5s
-Made 104869 iterations in 5s
-Made 111535 iterations in 5s
-Made 104705 iterations in 5s
+Made 23133 iterations in 00:00:01
+Made 81106 iterations in 00:00:01
+Made 104869 iterations in 00:00:01
+Made 111535 iterations in 00:00:01
+Made 104705 iterations in 00:00:01
 
         With type cache:
-Made 44492 iterations in 5s
-Made 284389 iterations in 5s
-Made 286152 iterations in 5s
-Made 276416 iterations in 5s
-Made 274822 iterations in 5s
+Made 44492 iterations in 00:00:01
+Made 284389 iterations in 00:00:01
+Made 286152 iterations in 00:00:01
+Made 276416 iterations in 00:00:01
+Made 274822 iterations in 00:00:01
+
+
+        On Dell box at the office:
+Made 373190 iterations in 00:00:01
+Made 394550 iterations in 00:00:01
+Made 426973 iterations in 00:00:01
+Made 436008 iterations in 00:00:01
+Made 434064 iterations in 00:00:01
 
 */
         [Test]
@@ -45,15 +56,16 @@ Made 274822 iterations in 5s
         {
             var iterations = 0L;
             var keepRunning = true;
+            var duration = TimeSpan.FromSeconds(1);
 
             Task.Run(async () =>
             {
-                await Task.Delay(TimeSpan.FromSeconds(1));
+                await Task.Delay(duration);
 
                 Volatile.Write(ref keepRunning, false);
             });
 
-            var headers = new Dictionary<string, string>{};
+            var headers = new Dictionary<string, string>();
             var message = new Message(headers, new RandomMessage("hello thereø how's it goin?"));
             var transportMessage = await _serializer.Serialize(message);
 
@@ -64,13 +76,91 @@ Made 274822 iterations in 5s
                 iterations++;
             }
 
-            Console.WriteLine($"Made {iterations} iterations in 5s");
+            Console.WriteLine($"Made {iterations} iterations in {duration}");
         }
+
+        [Test]
+        public async Task FormatTypeAsExpected_Default()
+        {
+            var expectedTypeName = typeof(SomeType).GetSimpleAssemblyQualifiedName();
+            var serializer = new JsonSerializer(new SimpleAssemblyQualifiedMessageTypeNameConvention());
+
+            var message = new Message(new Dictionary<string, string>(), new SomeType());
+            var transportMessage = await serializer.Serialize(message);
+
+            var type = transportMessage.Headers.GetValue(Headers.Type);
+
+            Console.WriteLine($@"
+
+Serialized type name: {type}
+  Expected type name: {expectedTypeName}
+
+");
+
+            var roundtrippedMessage = await serializer.Deserialize(transportMessage);
+
+            Assert.That(type, Is.EqualTo(expectedTypeName));
+            Assert.That(roundtrippedMessage.Body, Is.TypeOf<SomeType>());
+        }
+
+        [Test]
+        public async Task FormatTypeAsExpected_CustomWithFallback()
+        {
+            var serializer = new JsonSerializer(new CustomTypeNameConventionBuilder().AllowFallbackToDefaultConvention().GetConvention());
+
+            var expectedTypeName = typeof(SomeType).GetSimpleAssemblyQualifiedName();
+
+            var message = new Message(new Dictionary<string, string>(), new SomeType());
+            var transportMessage = await serializer.Serialize(message);
+
+            var type = transportMessage.Headers.GetValue(Headers.Type);
+
+            Console.WriteLine($@"
+
+Serialized type name: {type}
+  Expected type name: {expectedTypeName}
+
+");
+
+            var roundtrippedMessage = await serializer.Deserialize(transportMessage);
+
+            Assert.That(type, Is.EqualTo(expectedTypeName));
+            Assert.That(roundtrippedMessage.Body, Is.TypeOf<SomeType>());
+        }
+
+        [Test]
+        public async Task FormatTypeAsExpected_Custom()
+        {
+            var serializer = new JsonSerializer(new CustomTypeNameConventionBuilder()
+                .AddWithShortName<SomeType>()
+                .GetConvention());
+
+            const string expectedTypeName = "SomeType";
+
+            var message = new Message(new Dictionary<string, string>(), new SomeType());
+            var transportMessage = await serializer.Serialize(message);
+
+            var type = transportMessage.Headers.GetValue(Headers.Type);
+
+            Console.WriteLine($@"
+
+Serialized type name: {type}
+  Expected type name: {expectedTypeName}
+
+");
+
+            var roundtrippedMessage = await serializer.Deserialize(transportMessage);
+
+            Assert.That(type, Is.EqualTo(expectedTypeName));
+            Assert.That(roundtrippedMessage.Body, Is.TypeOf<SomeType>());
+        }
+
+        class SomeType { }
 
         [Test]
         public async Task WorksWithoutFullTypeNameHandlingToo()
         {
-            var simpleSerializer = new JsonSerializer(new JsonSerializerSettings { TypeNameHandling = TypeNameHandling.None });
+            var simpleSerializer = new JsonSerializer(new SimpleAssemblyQualifiedMessageTypeNameConvention(), new JsonSerializerSettings { TypeNameHandling = TypeNameHandling.None });
             var message = new RandomMessage("hei allihoppa");
             var transportMessage = await simpleSerializer.Serialize(new Message(new Dictionary<string, string>(), message));
             var roundtrippedMessage = (await simpleSerializer.Deserialize(transportMessage)).Body;