+++ Rebus/Pipeline/Send/AssignDefaultHeadersStep.cs
@@ -1,8 +1,8 @@
 ﻿using Rebus.Messages;
 using System;
 using System.Threading.Tasks;
+using Rebus.Serialization;
 using Rebus.Transport;
-using Rebus.Extensions;
 using Rebus.Time;
 
 namespace Rebus.Pipeline.Send
@@ -22,22 +22,24 @@ namespace Rebus.Pipeline.Send
 
 3) a 'rbs2-senttime' with the current time.
 
-4) 'rbs2-msg-type' with the message's simple assembly-qualified type name (*).
-
-(*) Unless explicitly set to something else")]
+-4) 'rbs2-msg-type' with the message's simple assembly-qualified type name (*).
+-
+-(*) Unless explicitly set to something else")]
     public class AssignDefaultHeadersStep : IOutgoingStep
     {
+        readonly IMessageTypeNameConvention _messageTypeNameConvention;
         readonly IRebusTime _rebusTime;
-        readonly bool _hasOwnAddress;
         readonly string _senderAddress;
         readonly string _returnAddress;
+        readonly bool _hasOwnAddress;
 
         /// <summary>
         /// Constructs the step, getting the input queue address from the given <see cref="ITransport"/>
         /// </summary>
-        public AssignDefaultHeadersStep(ITransport transport, IRebusTime rebusTime, string defaultReturnAddressOrNull)
+        public AssignDefaultHeadersStep(ITransport transport, IRebusTime rebusTime, IMessageTypeNameConvention messageTypeNameConvention, string defaultReturnAddressOrNull)
         {
             _rebusTime = rebusTime ?? throw new ArgumentNullException(nameof(rebusTime));
+            _messageTypeNameConvention = messageTypeNameConvention ?? throw new ArgumentNullException(nameof(messageTypeNameConvention));
             _senderAddress = transport.Address;
             _returnAddress = defaultReturnAddressOrNull ?? transport.Address;
             _hasOwnAddress = !string.IsNullOrWhiteSpace(_senderAddress);
@@ -50,7 +52,6 @@ namespace Rebus.Pipeline.Send
         {
             var message = context.Load<Message>();
             var headers = message.Headers;
-            var messageType = message.Body.GetType();
 
             if (!headers.ContainsKey(Headers.MessageId))
             {
@@ -59,7 +60,7 @@ namespace Rebus.Pipeline.Send
 
             if (_hasOwnAddress && !headers.ContainsKey(Headers.ReturnAddress))
             {
-                headers[Headers.ReturnAddress] =  _returnAddress;
+                headers[Headers.ReturnAddress] = _returnAddress;
             }
 
             headers[Headers.SentTime] = _rebusTime.Now.ToString("O");
@@ -75,7 +76,9 @@ namespace Rebus.Pipeline.Send
 
             if (!headers.ContainsKey(Headers.Type))
             {
-                headers[Headers.Type] = messageType.GetSimpleAssemblyQualifiedName();
+                var messageType = message.Body.GetType();
+
+                headers[Headers.Type] = _messageTypeNameConvention.GetTypeName(messageType);
             }
 
             await next();