+++ Rebus/Retry/FailFast/FailFastStep.cs
@@ -4,6 +4,10 @@ using Rebus.Bus;
 using Rebus.Exceptions;
 using Rebus.Messages;
 using Rebus.Pipeline;
+using Rebus.Retry.Simple;
+using Rebus.Transport;
+// ReSharper disable SuggestBaseTypeForParameter
+
 // ReSharper disable ArgumentsStyleLiteral
 
 namespace Rebus.Retry.FailFast
@@ -19,14 +23,18 @@ This allows the SimpleRetryStrategyStep to move it to the error queue.")]
     {
         readonly IErrorTracker _errorTracker;
         readonly IFailFastChecker _failFastChecker;
+        readonly IErrorHandler _errorHandler;
+        readonly ITransport _transport;
 
         /// <summary>
         /// Constructs the step, using the given error tracker
         /// </summary>
-        public FailFastStep(IErrorTracker errorTracker, IFailFastChecker failFastChecker)
+        public FailFastStep(IErrorTracker errorTracker, IFailFastChecker failFastChecker, IErrorHandler errorHandler, ITransport transport)
         {
             _errorTracker = errorTracker ?? throw new ArgumentNullException(nameof(errorTracker));
             _failFastChecker = failFastChecker ?? throw new ArgumentNullException(nameof(failFastChecker));
+            _errorHandler = errorHandler ?? throw new ArgumentNullException(nameof(errorHandler));
+            _transport = transport ?? throw new ArgumentNullException(nameof(transport));
         }
 
         /// <summary>
@@ -39,6 +47,13 @@ This allows the SimpleRetryStrategyStep to move it to the error queue.")]
             try
             {
                 await next();
+
+                var deadletterCommand = context.Load<ManualDeadletterCommand>();
+
+                if (deadletterCommand != null)
+                {
+                    await ProcessDeadletterCommand(context, deadletterCommand);
+                }
             }
             catch (Exception exception)
             {
@@ -46,10 +61,30 @@ This allows the SimpleRetryStrategyStep to move it to the error queue.")]
                 var messageId = transportMessage.GetMessageId();
                 if (_failFastChecker.ShouldFailFast(messageId, exception))
                 {
-                    _errorTracker.MarkAsFinal(messageId);
+                    // if we're currently executing a 2nd level retry, it's the 2nd level surrogate message ID we must mark as final
+                    var messageIdToMarkAsFinal = context.Load<bool>(SimpleRetryStrategyStep.DispatchAsFailedMessageKey)
+                        ? SimpleRetryStrategyStep.GetSecondLevelMessageId(messageId)
+                        : messageId;
+
+                    _errorTracker.MarkAsFinal(messageIdToMarkAsFinal);
                 }
                 throw;
             }
+        }
+
+        async Task ProcessDeadletterCommand(IncomingStepContext context, ManualDeadletterCommand deadletterCommand)
+        {
+            var originalTransportMessage = context.Load<OriginalTransportMessage>() ?? throw new RebusApplicationException("Could not find the original transport message in the current incoming step context");
+
+            var transportMessage = originalTransportMessage.TransportMessage.Clone();
+            var errorDetails = deadletterCommand.ErrorDetails ?? "Manually dead-lettered";
+
+            transportMessage.Headers[Headers.ErrorDetails] = errorDetails;
+            transportMessage.Headers[Headers.SourceQueue] = _transport.Address;
+
+            var transactionContext = context.Load<ITransactionContext>();
+
+            await _errorHandler.HandlePoisonMessage(transportMessage, transactionContext, new RebusApplicationException(errorDetails));
         }
     }
 }