+++ Rebus/Serialization/Json/NewtonsoftJsonConfigurationExtensions.cs
@@ -19,7 +19,7 @@ namespace Rebus.Serialization.Json
         public static void UseNewtonsoftJson(this StandardConfigurer<ISerializer> configurer)
         {
             if (configurer == null) throw new ArgumentNullException(nameof(configurer));
-            var settings = new JsonSerializerSettings {TypeNameHandling = TypeNameHandling.All};
+            var settings = new JsonSerializerSettings { TypeNameHandling = TypeNameHandling.All };
 
             RegisterSerializer(configurer, settings, Encoding.UTF8);
         }
@@ -35,20 +35,7 @@ namespace Rebus.Serialization.Json
         {
             if (configurer == null) throw new ArgumentNullException(nameof(configurer));
 
-            TypeNameHandling GetTypeNameHandling()
-            {
-                switch (mode)
-                {
-                    case JsonInteroperabilityMode.FullTypeInformation:
-                        return TypeNameHandling.All;
-                    case JsonInteroperabilityMode.PureJson:
-                        return TypeNameHandling.None;
-                    default:
-                        throw new ArgumentOutOfRangeException(nameof(mode), mode, $"Unknown {typeof(JsonInteroperabilityMode).Name} value");
-                }
-            }
-
-            var settings = new JsonSerializerSettings {TypeNameHandling = GetTypeNameHandling()};
+            var settings = new JsonSerializerSettings { TypeNameHandling = GetTypeNameHandling(mode) };
 
             RegisterSerializer(configurer, settings, Encoding.UTF8);
         }
@@ -62,14 +49,27 @@ namespace Rebus.Serialization.Json
             if (configurer == null) throw new ArgumentNullException(nameof(configurer));
             if (settings == null) throw new ArgumentNullException(nameof(settings));
 
-            RegisterSerializer(configurer, settings, encoding);
+            RegisterSerializer(configurer, settings, encoding ?? Encoding.UTF8);
+        }
+
+        static TypeNameHandling GetTypeNameHandling(JsonInteroperabilityMode mode)
+        {
+            switch (mode)
+            {
+                case JsonInteroperabilityMode.FullTypeInformation:
+                    return TypeNameHandling.All;
+                case JsonInteroperabilityMode.PureJson:
+                    return TypeNameHandling.None;
+                default:
+                    throw new ArgumentOutOfRangeException(nameof(mode), mode, $"Unknown {typeof(JsonInteroperabilityMode).Name} value");
+            }
         }
 
         static void RegisterSerializer(StandardConfigurer<ISerializer> configurer, JsonSerializerSettings settings, Encoding encoding)
         {
             if (configurer == null) throw new ArgumentNullException(nameof(configurer));
 
-            configurer.Register(c => new JsonSerializer(settings, encoding ?? Encoding.UTF8));
+            configurer.Register(c => new JsonSerializer(c.Get<IMessageTypeNameConvention>(), settings, encoding));
         }
     }
 }